c: fix the warnings that flag real defects
krun printed its 64-bit counters with %lu, which is correct on LP64
platforms only by coincidence and undefined on LLP64 Windows, where
unsigned long is 32 bits -- the pixel and timing numbers there were
garbage; print with PRIu64. The POSIX UDP layer did void-pointer
arithmetic (a GNU extension; its Windows twin already casts to char*).
ktemp carried an unreachable epilogue after its run-forever loop.
Narrowing conversions are now explicit where they are intended: the
test-pulse period byte is masked like its neighboring lines, and the
C++ chrono accessors cast their counts to the declared types. Two
maybe-uninitialized warnings, both false positives by inspection, are
settled by initializing at declaration.

+++ cxx/include/katherinexx/acquisition.hpp
@@ -100,7 +100,7 @@ public:
     {
         using namespace std::chrono;
 
-        int res = katherine_acquisition_init(&acq_, dev.c_dev(), reinterpret_cast<void *>(this), md_buffer_size, pixel_buffer_size, duration_cast<milliseconds>(report_timeout).count(), duration_cast<milliseconds>(fail_timeout).count());
+        int res = katherine_acquisition_init(&acq_, dev.c_dev(), reinterpret_cast<void *>(this), md_buffer_size, pixel_buffer_size, static_cast<int>(duration_cast<milliseconds>(report_timeout).count()), static_cast<int>(duration_cast<milliseconds>(fail_timeout).count()));
         if (res != 0) {
             throw katherine::system_error{res};
         }
